refactor continues
