+++ scripts/python/figures/OA_regression.py
@@ -1,8 +1,11 @@
-# %%
 from typing import List
 import pandas as pd
 from pathlib import Path
 import numpy as np
+import statsmodels.api as sm
+from patsy import dmatrices
+import seaborn as sns
+from matplotlib import pyplot as plt
 
 data_path = Path(__file__).parents[3] / "data" / "correlates"
 
@@ -45,7 +48,6 @@ def parse_dfs(directory: Path) -> List[pd.DataFrame]:
     return combined_df
 
 
-# %%
 df = parse_dfs(data_path)
 
 poverty_using = pd.read_csv(
@@ -150,13 +152,6 @@ election = election.drop(columns={"year"})
 
 df = df.merge(election, on="FIPS", how="left")
 
-# %%
-import statsmodels.api as sm
-from patsy import dmatrices
-import seaborn as sns
-from matplotlib import pyplot as plt
-
-
 dfs = [
     dem_only_using,
     poverty_using,
@@ -191,17 +186,6 @@ for name in names:
         coef = model_res.params[1]
         pvalue = model_res.pvalues[1]
         std_error = model_res.HC1_se[1]
-        # g = sns.jointplot(
-        #     data=tfi,
-        #     x=col,
-        #     y=f"selection_ratio_log_{name}",
-        #     kind="reg",
-        # )
-        # plt.text(x=0.5, y=2.2, s=f"slope: {coef:.3f}\n ci: {pvalue:.3f}")
-        # g.ax_joint.set_ylabel("log(selection_ratio)")
-        # g.ax_joint.set_xlabel(col)
-        # plt.title(f"{name} {col}")
-        # plt.show()
         result = f"{coef:.3f} ({std_error:.3f})"
         if pvalue <= 0.05:
             result += "*"
@@ -212,22 +196,6 @@ for name in names:
         results += [[name, col, result]]
 
 result_df = pd.DataFrame(results, columns=["model", "correlate", "coef (p-value)"])
-
-
-def check_coef(df, col1, col2):
-    df = df[~df[col1].isnull()]
-    df = df[~df[col2].isnull()]
-    y, X = dmatrices(f"{col1} ~ {col2}", data=df, return_type="dataframe")
-    model = sm.OLS(
-        y,
-        X,
-    )
-    model_res = model.fit()
-    model_res = model_res.get_robustcov_results(cov_type="HC1")
-    print(model_res.summary())
-
-
-# %%
 
 correlate_order = [
     "bwratio",
@@ -288,113 +256,5 @@ pivot_df = pivot_df.loc[correlate_order]
 pivot_df = pivot_df.rename(name_conv, axis=0)
 pivot_df = pivot_df.rename(model_conv, axis=1)
 
-# %%
-
-pivot_df.to_csv(data_path / "correlate_pivot.csv")
-df.to_csv(data_path / "processed_correlates.csv")
-
-# %%
-
-legal_states = {
-    "08": 2012,
-    "25": 2016,
-    "26": 2018,
-    "41": 2014,
-    "50": 2013,
-    "53": 2012,
-}
-
-
-def reporting_throughout(df: pd.DataFrame):
-    reported = (
-        (df.groupby("FIPS").size() == df.year.nunique())
-        .to_frame("reported")
-        .reset_index()
-    )
-    df = df.merge(reported, on="FIPS")
-    return df[df.reported]
-
-
-def filter_legal(df: pd.DataFrame, legal: bool = False):
-    def _remove_legal(key):
-        year = legal_states[key]
-        condition = (df.year >= year) & (df.FIPS.apply(lambda x: x.startswith(key)))
-        if not legal:
-            return df[~condition]
-        else:
-            return df[condition]
-
-    if not legal:
-        for key in legal_states.keys():
-            df = _remove_legal(key)
-        return df
-    else:
-        df_temp = pd.DataFrame()
-        for key in legal_states.keys():
-            df_temp = df_temp.append(_remove_legal(key))
-        return df_temp
-
-
-def get_year_data(name: str, log_ratio: bool, legal: bool, reported: bool):
-    filename = "selection_ratio_county_2010-2019_bootstraps_1000"
-    if name == "Dem only":
-        filename += ".csv"
-    if name == "Dem + Pov":
-        filename += "_poverty.csv"
-    if name == "Dem + metro":
-        filename += "_metro.csv"
-    if name == "Buying":
-        filename += "_poverty_buying.csv"
-    if name == "Buying Outside":
-        filename += "_poverty_buying_outside.csv"
-    if name == "Arrests":
-        filename += "_poverty_arrests.csv"
-    df = pd.read_csv(
-        data_path.parent.parent / "turing_output" / filename, dtype={"FIPS": str}
-    )
-    if legal:
-        df = filter_legal(df, legal=True)
-    else:
-        df = filter_legal(df, legal=False)
-    if reported:
-        df = reporting_throughout(df)
-    if log_ratio:
-        df["selection_ratio"] = np.log(df["selection_ratio"])
-    y, X = dmatrices(f"selection_ratio ~ year", data=df, return_type="dataframe")
-    model = sm.WLS(
-        y,
-        X,
-        weights=1 / df["var_log"],
-    )
-    model_res = model.fit()
-    model_res = model_res.get_robustcov_results(cov_type="HC1")
-    coef = model_res.params[1]
-    pvalue = model_res.pvalues[1]
-    std_error = model_res.HC1_se[1]
-    result = f"{coef:.3f} ({std_error:.3f})"
-    if pvalue <= 0.05:
-        result += "*"
-    if pvalue <= 0.01:
-        result += "*"
-    if pvalue <= 0.001:
-        result += "*"
-    return result
-
-
-time_names = [f"Time {i}" for i in range(1, 5)] + [
-    f"Time {i} Log SR" for i in range(1, 5)
-]
-
-time_results = [
-    [get_year_data(name, False, False, True) for name in model_names],
-    [get_year_data(name, False, False, False) for name in model_names],
-    [get_year_data(name, False, True, True) for name in model_names],
-    [get_year_data(name, False, True, False) for name in model_names],
-    [get_year_data(name, True, False, True) for name in model_names],
-    [get_year_data(name, True, False, False) for name in model_names],
-    [get_year_data(name, True, True, True) for name in model_names],
-    [get_year_data(name, True, True, False) for name in model_names],
-]
-
-time_df = pd.DataFrame(time_results, columns=model_names, index=time_names)
-# %%
+pivot_df.to_csv(data_path / "OA_coef_pivot.csv")
+df.to_csv(data_path / "OA_processed.csv")
